--- src/app/service/posts.ts
@@ -1,9 +1,0 @@
-import { promises as fs } from 'fs';
-import path from 'path';
-
-import { Menu } from '@/lib/types';
-
-export function getAllMenus(): Promise<Menu[]> {
-  const filePath = path.join(process.cwd(), 'data', 'menus.json');
-  return fs.readFile(filePath, 'utf-8').then<Menu[]>(JSON.parse);
-}
