--- src/lib/types.d.ts
@@ -1,3 +1,0 @@
-/// <reference types="gtag.js" />
-
-declare module 'gtag.js';
