+++ src/types/Heading.ts
@@ -1,0 +1,6 @@
+export type Heading = {
+  text: string;
+  flag: string;
+  parent?: string;
+  isSub: boolean;
+};
